Figured out bug in windows version, seekg was getting a long (4 bits) instead of a long long (8 bits) limiting file size to 2GB. I guess gcc corrects that stuff for you?

+++ src/bitreader.cpp
@@ -152,7 +152,7 @@ STRING bitreader::readLengthNSTRING(int N) {
   return out;
 }
 
-void bitreader::seekToByte(long g) {
+void bitreader::seekToByte(long long g) {
   input->seekg(g);
 }
